Flatten AutoRespawn
