+++ src/main/java/com/programmerdan/minecraft/simpleadminhacks/hacks/basic/AutoRespawn.java
@@ -1,4 +1,4 @@
-package com.programmerdan.minecraft.simpleadminhacks.hacks.basic.AutoRespawn;
+package com.programmerdan.minecraft.simpleadminhacks.hacks.basic;
 
 import com.google.common.base.Strings;
 import com.programmerdan.minecraft.simpleadminhacks.SimpleAdminHacks;
@@ -9,7 +9,11 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Consumer;
 import org.bukkit.Bukkit;
+import org.bukkit.boss.BarColor;
+import org.bukkit.boss.BarStyle;
+import org.bukkit.boss.BossBar;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -17,12 +21,14 @@ import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.event.player.PlayerRespawnEvent;
+import org.bukkit.scheduler.BukkitRunnable;
+import org.bukkit.scheduler.BukkitTask;
 import vg.civcraft.mc.civmodcore.util.Iteration;
 import vg.civcraft.mc.civmodcore.util.TextUtil;
 
 public final class AutoRespawn extends BasicHack {
 
-	private final Map<Player, _RespawnTimer> respawnTimers = new HashMap<>();
+	private final Map<Player, RespawnTimer> respawnTimers = new HashMap<>();
 
 	@AutoLoad
 	private long respawnDelay;
@@ -62,7 +68,7 @@ public final class AutoRespawn extends BasicHack {
 			plugin().info("Player [" + player.getName() + "] died, " +
 					"setting respawn timer: " + this.respawnDelay);
 			this.respawnTimers.computeIfPresent(player, (_player, timer) -> timer.stop());
-			this.respawnTimers.put(player, new _RespawnTimer(this, player,
+			this.respawnTimers.put(player, new RespawnTimer(this, player,
 					this.respawnDelay, this::autoRespawnPlayer));
 		}
 	}
@@ -83,7 +89,7 @@ public final class AutoRespawn extends BasicHack {
 			plugin().info("Player [" + player.getName() + "] logged in while dead, " +
 					"setting respawn timer: " + this.loginRespawnDelay);
 			this.respawnTimers.computeIfPresent(player, (_player, timer) -> timer.stop());
-			this.respawnTimers.put(player, new _RespawnTimer(this, player,
+			this.respawnTimers.put(player, new RespawnTimer(this, player,
 					this.loginRespawnDelay, this::autoRespawnPlayer));
 		}
 	}
@@ -114,5 +120,85 @@ public final class AutoRespawn extends BasicHack {
 	public static BasicHackConfig generate(final SimpleAdminHacks plugin, final ConfigurationSection config) {
 		return new BasicHackConfig(plugin, config);
 	}
-	
+
+	// ------------------------------------------------------------
+	// Respawn Timer
+	// ------------------------------------------------------------
+
+	private static final class RespawnTimer {
+
+		private final Consumer<Player> handler;
+		private BossBar bar;
+		private long previousTime;
+		private final long setTime;
+		private long timeRemaining;
+		private long secondTimer;
+		private BukkitTask processor;
+
+		RespawnTimer(final AutoRespawn hack, final Player player, final long delay, final Consumer<Player> handler) {
+			this.handler = handler;
+			this.previousTime = System.currentTimeMillis();
+			this.setTime = this.timeRemaining = delay;
+			this.secondTimer = 1000L;
+			this.bar = Bukkit.createBossBar(generateBarTitle(), BarColor.WHITE, BarStyle.SOLID);
+			this.bar.setVisible(true);
+			this.bar.setProgress(1.0d);
+			this.bar.addPlayer(player);
+			this.processor = new BukkitRunnable() {
+				@Override
+				public void run() {
+					tick();
+				}
+			}.runTaskTimer(hack.plugin(), 1L, 1L);
+		}
+
+		private String generateBarTitle() {
+			if (this.timeRemaining <= 1000L) {
+				return "Respawning now.";
+			}
+			if (this.timeRemaining < 60000L) {
+				return "Respawning in " + (this.timeRemaining / 1000L) + " seconds.";
+			}
+			if (this.timeRemaining < 120000L) {
+				return "Respawning in 1 minute.";
+			}
+			if (this.timeRemaining < 3600000L) {
+				return "Respawning in " + (this.timeRemaining / 60000L) + " minutes.";
+			}
+			return "Respawning in " + (this.timeRemaining / 3600000L) + " hours.";
+		}
+
+		private void tick() {
+			long currentTime = System.currentTimeMillis();
+			long timeDifference = currentTime - this.previousTime;
+			this.previousTime = currentTime;
+			this.timeRemaining -= timeDifference;
+			this.secondTimer -= timeDifference;
+			if (this.secondTimer > 0) {
+				return;
+			}
+			this.secondTimer = 1000L;
+			this.bar.setTitle(generateBarTitle());
+			this.bar.setProgress(Math.max((double) (this.timeRemaining - 1000L) / (double) this.setTime, 0));
+			if (this.timeRemaining > 0) {
+				return;
+			}
+			this.bar.getPlayers().forEach(this.handler);
+		}
+
+		public RespawnTimer stop() {
+			if (this.bar != null) {
+				this.bar.setVisible(false);
+				this.bar.removeAll();
+				this.bar = null;
+			}
+			if (this.processor != null) {
+				this.processor.cancel();
+				this.processor = null;
+			}
+			return null;
+		}
+
+	}
+
 }
