+++ src/main/resources/config.yml
@@ -9,21 +9,6 @@ hacks:
   ############################################################
   ## Basic Hacks
   ############################################################
-  AutoRespawn:
-    enabled: true
-    # Delay in MS
-    # Default: 300000 (five minutes)
-    respawnDelay: 300000
-    # The respawn delay (in MS) if someone logs in dead
-    # Default 0 (instant)
-    loginRespawnDelay: 0
-    # Jokey joke messages sent to the player if auto-respawned
-    # Deactivate by removing all the quotes below
-    respawnQuotes:
-      - <gold>Wait... how did I get here?
-      - <gold>Must've been a rough night o.o
-      - <gold>Why does my entire body ache?
-      - <gold>That <it>really<gold> hurt!
   AntiFastBreak:
     enabled: true
     breakDenyDuration: 3000
@@ -53,6 +38,21 @@ hacks:
       - SALMON
       - TURTLE
       - DOLPHIN
+  AutoRespawn:
+    enabled: true
+    # Delay in MS
+    # Default: 300000 (five minutes)
+    respawnDelay: 300000
+    # The respawn delay (in MS) if someone logs in dead
+    # Default 0 (instant)
+    loginRespawnDelay: 0
+    # Jokey joke messages sent to the player if auto-respawned
+    # Deactivate by removing all the quotes below
+    respawnQuotes:
+      - <gold>Wait... how did I get here?
+      - <gold>Must've been a rough night o.o
+      - <gold>Why does my entire body ache?
+      - <gold>That <it>really<gold> hurt!
   BuffSpanker:
     enabled: true
     # The list of PotionEffectType's that you do not wish players to receive.
